+++ library/docs/changelog.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `BlockingLaboratory` class that can read and write feature flags via blocking API.
+- `blocking()` function to `Laboratory` class that is an entry point to the blocking API.
+
+### Deprecated
+- All blocking functions on the `Laboratory` class. `BlockingLaboratory` available via `blocking()` function should be used instead.
+
 ## [0.9.1] - 2020-11-12
 
 ### Added